Add crate `git2` as `git`
To get a list of the branches, it's very easy to just do one of these:
- Look at `.git/refs/heads/`
- Run `git branch --format='%(refname:short)'`

But I plan on implementing some complex features later on and I think this crate will save me from a lot of headaches.

NOTE: There's a `git` crate, but it's deprecated in favor of `git2`.
      I don't think this renaming will hurt.