Set `~/` as the pre-selected folder
This prevents the user from clicking "Select" without a selected folder.
This only happened when you opened the dialog and clicked the button right away.
Now it will always have selected something.
(Well, I don't know if you can deselect a folder)

This wasn't a big of a deal, but it bothered me.
Clicking "Select" and just closing the dialog isn't actually selecting anything, is it?