Validate if selected folder is a repo
When you select an folder, it updates the window subtitle.
Any folder could be selected, not just repos.
You can't open another folder because the "Open" button would be gone.
The app thinks you opened an actual repo.

To validate, I check whether `git2` can open the folder as a repo.
If it can't, then it's not a repo and nothing happens.
You can try again.