Track presented view with enum instead of boolean
`content` can present multiple views depending on what you do with the sidebar.
It was using `repository_was_selected` to track which view to show.
It works fine for now, as it's the only bool needed, but with more views it could get bad.
I replaced it with an enum to say which view should be presented.
The `repository_was_selected` bool is no longer needed and was removed.

This approach reads better and it's also scalable.
A win-win for me.
I'm planing on more complex features, after all.