Dispatch repo as `PathBuf` instead of as GFile
To list branches, `app` needs the repo selected from the file chooser in `content`.
Then `content` dispatches the repo's folder as a `gio::File` to `app`.
The repo is opened with `git::Repository::open(_)`, which takes a path.

In `app`, the `gio::File` was only used to get the `PathBuf` from it.
Since `content` already had the path to validate the repo, why not just send that instead?