Prevent list background from stretching too much
Lists were always occupying the whole view, even when they just had a few items.
It's like it had the last item gigantic, but its highlight would have the correct size.
The problem then is the background of the list.
The items are okay and so are overflowing lists.

I tried combinations of `set_vexpand: false` in the hierarchy, to no avail.
Using `adw::StatusPage` solved the background issue, but centered small lists vertically.
I decided to just put the list in a `gtk::Box`.
Works just fine for all list sizes now.